+++ api/src/main/java/fr/funixgaming/api/client/external_api_impl/twitch/reference/dtos/requests/TwitchChatAnnouncement.java
@@ -1,4 +1,4 @@
-package fr.funixgaming.api.server.external_api_impl.twitch.reference.dtos.requests;
+package fr.funixgaming.api.client.external_api_impl.twitch.reference.dtos.requests;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
