+++ api/src/main/java/fr/funixgaming/api/client/external_api_impl/twitch/reference/dtos/responses/channel/video/TwitchChannelClipCreationDTO.java
@@ -1,4 +1,4 @@
-package fr.funixgaming.api.server.external_api_impl.twitch.reference.dtos.responses.channel.video;
+package fr.funixgaming.api.client.external_api_impl.twitch.reference.dtos.responses.channel.video;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
