+++ service/src/main/java/fr/funixgaming/api/server/external_api_impl/twitch/reference/clients/channel/TwitchReferenceChannelClient.java
@@ -1,10 +1,10 @@
 package fr.funixgaming.api.server.external_api_impl.twitch.reference.clients.channel;
 
+import fr.funixgaming.api.client.external_api_impl.twitch.reference.dtos.requests.TwitchChannelUpdateDTO;
+import fr.funixgaming.api.client.external_api_impl.twitch.reference.dtos.responses.TwitchDataResponseDTO;
+import fr.funixgaming.api.client.external_api_impl.twitch.reference.dtos.responses.channel.TwitchChannelDTO;
+import fr.funixgaming.api.client.external_api_impl.twitch.reference.dtos.responses.channel.chat.TwitchChannelUserDTO;
 import fr.funixgaming.api.server.external_api_impl.twitch.reference.clients.TwitchReferenceRequestInterceptor;
-import fr.funixgaming.api.server.external_api_impl.twitch.reference.dtos.requests.TwitchChannelUpdateDTO;
-import fr.funixgaming.api.server.external_api_impl.twitch.reference.dtos.responses.TwitchDataResponseDTO;
-import fr.funixgaming.api.server.external_api_impl.twitch.reference.dtos.responses.channel.TwitchChannelDTO;
-import fr.funixgaming.api.server.external_api_impl.twitch.reference.dtos.responses.channel.chat.TwitchChannelUserDTO;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.bind.annotation.*;
