+++ service/src/main/java/fr/funixgaming/api/server/external_api_impl/twitch/reference/clients/chat/TwitchReferenceChatClient.java
@@ -1,9 +1,9 @@
 package fr.funixgaming.api.server.external_api_impl.twitch.reference.clients.chat;
 
+import fr.funixgaming.api.client.external_api_impl.twitch.reference.dtos.requests.TwitchChatAnnouncement;
+import fr.funixgaming.api.client.external_api_impl.twitch.reference.dtos.responses.TwitchDataResponseDTO;
+import fr.funixgaming.api.client.external_api_impl.twitch.reference.dtos.responses.channel.chat.TwitchChannelChattersDTO;
 import fr.funixgaming.api.server.external_api_impl.twitch.reference.clients.TwitchReferenceRequestInterceptor;
-import fr.funixgaming.api.server.external_api_impl.twitch.reference.dtos.requests.TwitchChatAnnouncement;
-import fr.funixgaming.api.server.external_api_impl.twitch.reference.dtos.responses.TwitchDataResponseDTO;
-import fr.funixgaming.api.server.external_api_impl.twitch.reference.dtos.responses.channel.chat.TwitchChannelChattersDTO;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.bind.annotation.*;
